requirements.txt
torch == 1.6.0
torchtext == 0.3.1
discord.py == 1.5.1
asyncio == 3.4.3
